deployment/ECS-Catboost/requirements.txt
flask==3.1.0
boto3==1.35.43
numpy<2.0
gunicorn==23.0.0
scikit-learn==1.5.2
python-dotenv==1.0.1
flask_cors==5.0.1
catboost==1.2.7
